requirements.txt
langchain
langchain-openai
langchain-groq
langchain-google-genai
python-dotenv
pydantic
tiktoken
pytest
